+++ src/pass/std_nist.rs
@@ -8,7 +8,7 @@ pub const DEFAULT_NORMALIZATION: Normalization = Normalization::Nfkc;
 pub const DEFAULT_PASSWORD_MIN_LEN: usize = 8;
 pub const DEFAULT_PASSWORD_MAX_LEN: usize = 128;
 pub const DEFAULT_ALGORITHM: Algorithm = Algorithm::Pbkdf2;
-pub const DEFAULT_LENGTH_CALCULATION: LengthCalculationMethod = LengthCalculationMethod::Characters;
+pub const DEFAULT_LENGTH_CALCULATION: LengthCalculationMethod = LengthCalculationMethod::CodePoints;
 pub const DEFAULT_SALT_LEN: usize = 16; // In bytes
 pub const DEFAULT_XHMAC_ALGORITHM: HashFunction = HashFunction::Sha512;
 pub const MIN_SALT_LEN: usize = 4; // In bytes
@@ -18,11 +18,8 @@ pub const NB_ITER_MIN: u32 = 10_000;
 
 pub fn is_valid(h: &HashBuilder) -> bool {
 	// Length calculation
-	match h.length_calculation {
-		LengthCalculationMethod::Characters => {}
-		LengthCalculationMethod::Bytes => {
-			return false;
-		}
+	if h.length_calculation != LengthCalculationMethod::CodePoints {
+		return false;
 	}
 
 	// Salt length.